rename types.h -> fpc-types.h
